notice_communication/requirements.txt
pydub==0.25.1
google-cloud-texttospeech==2.9.0
requests-oauthlib==1.3.0
tweepy==4.4.0
